Support restoring from remote PostgreSQL server
This allows the restoration of DB data to be done from a remote
PostgreSQL server.  Unlike backup, which requires only access to the
database being backed up, restoration requires higher privileges: it
needs to be able to create users and databases.

Local restoration uses the PostgreSQL administrative account
"postgres".  Remote servers require a suitably-powerful PostgreSQL
username and password to be supplied (via DB_ADMIN_USER and
DB_ADMIN_PASSWORD) to allow pg_restore to do its work.

Note that backup.sh already supports local and remote backup via
appropriate setting of DB_HOST and DB_PORT.

Also note that DB_ADMIN_DB is required because we need to connect to a
database other than the one we are restoring because we can't drop
DB_NAME or delete DB_USER if we are connected to the DB_NAME database.