Support command line arguments
Allow the backup directory to be specified via the command line, if
desired.  If a backups directory is not specified on the command line,
it will fall back on the default BACKUP_ROOT set in the script.

We also ensure that a directory from which to restore is specified as
an argument to restore.sh, as we have no default for that.