Handle PostgreSQL logs
If the PostgreSQL server is not running locally on the system being
backed up/restored then we need to avoid backing up/restoring the
PostgreSQL server logs because there won't be any on this system.

Note that DB_IS_REMOTE refers to whether the PostgreSQL server is
installed remotely or locally to this system.  It does not strictly
refer to the connection transport used.